requirements.txt
asgiref==3.6.0
Django==4.1.7
django-phonenumber-field==7.0.2
gunicorn==20.1.0
phonenumberslite==8.13.6
psycopg2==2.9.5
sqlparse==0.4.3
